requirements.txt
numpy
dacite
pyyaml
cvxpy
matplotlib
ecos